R3 - MODALS_CONSTENT moved to utils/constants
