--- src/utils/constants.ts
@@ -1,5 +1,0 @@
-export const LOADING_STATUS = {
-    IDLE: 'idle',
-    LOADING: 'loading',
-    FAILED: 'failed',
-};
