+++ src/utils/constants.tsx
@@ -1,0 +1,20 @@
+export const LOADING_STATUS = {
+    IDLE: 'idle',
+    LOADING: 'loading',
+    FAILED: 'failed',
+};
+
+export const MODALS_CONTENT = {
+    info: {
+        title: 'Information',
+        content: <p>Information modal content</p>,
+    },
+    stats: {
+        title: 'Statistics',
+        content: <p>Statistics modal content</p>,
+    },
+    settings: {
+        title: 'Settings',
+        content: <p>Settings modal content</p>,
+    },
+};
